Firmware/PythonCodes/requirements.txt
colorama==0.4.6
numpy==2.4.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
pyqtgraph==0.14.0
pyserial==3.5
